move bucket iterators to separate files
