+++ src/bucket/BucketApplicator.h
@@ -5,6 +5,7 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "bucket/Bucket.h"
+#include "bucket/BucketInputIterator.h"
 #include "database/Database.h"
 #include "util/XDRStream.h"
 #include <memory>
@@ -21,7 +22,7 @@ class Database;
 class BucketApplicator
 {
     Database& mDb;
-    Bucket::InputIterator mBucketIter;
+    BucketInputIterator mBucketIter;
     size_t mSize{0};
 
   public:
