+++ src/invariant/BucketListIsConsistentWithDatabase.cpp
@@ -4,6 +4,7 @@
 
 #include "invariant/BucketListIsConsistentWithDatabase.h"
 #include "bucket/Bucket.h"
+#include "bucket/BucketInputIterator.h"
 #include "crypto/Hex.h"
 #include "database/Database.h"
 #include "invariant/InvariantManager.h"
@@ -46,7 +47,7 @@ BucketListIsConsistentWithDatabase::checkOnBucketApply(
     uint64_t nAccounts = 0, nTrustLines = 0, nOffers = 0, nData = 0;
     bool hasPreviousEntry = false;
     BucketEntry previousEntry;
-    for (Bucket::InputIterator iter(bucket); iter; ++iter)
+    for (BucketInputIterator iter(bucket); iter; ++iter)
     {
         auto const& e = *iter;
         if (hasPreviousEntry && !BucketEntryIdCmp{}(previousEntry, e))
